ENH: create a dataset of pre-registered motors. See #664 (#744)
* ENH: Create a dataset of pre-registered motors. See #664

I followed the recommendation "Download and save several .eng files in the repo so we can install it along with the rocketpy package itself".

The website thrustcurve.org was very useful to search for some .eng files. I mainly focused in some of the main brands on the market: Cesaroni, Aero Tech, Animal Motors and Loki. And also focused on classes K to M, because this is the main range of total impulse that I'm used to seeing in rocketry. I tried to pick motors with a difference of about 300~600Ns in total impulse.

Some more improvements than can also be made following this issue are expanding the dataset for whole SolidMotor objects, more than only thrust curves. I think this would be what the recommendation "Save .json files with all the information we may find available on internet" could mean. I decided to go for the simple for now, but having the thrust curves is a good first step to implementing that in the future, which I would totally be able to do!

* Update CHANGELOG.md

* MNT: git rename motor eng files in data folder

* ENH: Create a dataset of pre-registered motors. See #664

I followed the recommendation "Download and save several .eng files in the repo so we can install it along with the rocketpy package itself".

The website thrustcurve.org was very useful to search for some .eng files. I mainly focused in some of the main brands on the market: Cesaroni, Aero Tech, Animal Motors and Loki. And also focused on classes K to M, because this is the main range of total impulse that I'm used to seeing in rocketry. I tried to pick motors with a difference of about 300~600Ns in total impulse.

Some more improvements than can also be made following this issue are expanding the dataset for whole SolidMotor objects, more than only thrust curves. I think this would be what the recommendation "Save .json files with all the information we may find available on internet" could mean. I decided to go for the simple for now, but having the thrust curves is a good first step to implementing that in the future, which I would totally be able to do!

Update CHANGELOG.md

MNT: git rename motor eng files in data folder

---------

Co-authored-by: Gui-FernandesBR <guilherme_fernandes@usp.br>