ENH: Add Eccentricity to Stochastic Simulations (#792)
* wind factor bug corrected

the wind factor wasn't applied to the env.wind_velocity properties

* BUG: StochasticModel visualize attributes of a uniform distribution

It showed the nominal and the standard deviation values and it doesn't make sense in a uniform distribution. In a np.random.uniform the 'nominal value' is the lower bound of the distribution, and the 'standard deviation' value is the upper bound. Now, a new condition has been added for the uniform distributions where the mean and semi range are calculated and showed. This way the visualize_attribute function will show the whole range where the random values are uniformly taken in

* variable names corrections

* Corrections requested by the pylint test

* ENH: more intuitive uniform distribution display in StochasticModel

Co-authored-by: MateusStano <69485049+MateusStano@users.noreply.github.com>

* ENH: Eccentricities added to the StochasticRocket class

A bug has been corrected in Flight class and an enhancement has been performed in the Rocket class as well

* BUG: thrust eccentricity bug corrected

eccentricity_y was defined by x coordinate and eccentricity_x was defined by y coordinate

* BUG: Undo some Rocket class changes

* ENH: add eccentricities to StochasticRocket

* BUG: fix MonteCarlo eccentricity inputs

* ENH: pylint and ruff recommended changes

* TST: fix tests with eccentricity

---------

Co-authored-by: Gui-FernandesBR <guilherme_fernandes@usp.br>