ENH: Introduce Net Thrust with pressure corrections (#789)
* wind factor bug corrected

the wind factor wasn't applied to the env.wind_velocity properties

* BUG: StochasticModel visualize attributes of a uniform distribution

It showed the nominal and the standard deviation values and it doesn't make sense in a uniform distribution. In a np.random.uniform the 'nominal value' is the lower bound of the distribution, and the 'standard deviation' value is the upper bound. Now, a new condition has been added for the uniform distributions where the mean and semi range are calculated and showed. This way the visualize_attribute function will show the whole range where the random values are uniformly taken in

* variable names corrections

* Corrections requested by the pylint test

* ENH: Add pressure corrections for thrust in SolidMotor

The thrust generated by a SolidMotor is now adjusted for the atmospheric pressure. To achieve that, a new attribute, 'vacuum_thrust', has been created. The 'net_thrust' is the result of 'vacuum_thrust' minus the atmospheric pressure multiplied by the nozzle area.

* ENH: pylint recommendations done

* ENH: net thrust method extended to the rest of the motor classes

* BUG: __post_processed_variables inconsistent array

* ENH: ruff reformatting

* Update rocketpy/motors/motor.py

Co-authored-by: Gui-FernandesBR <63590233+Gui-FernandesBR@users.noreply.github.com>

* ENH: Avoid breaking change

* ENH: Pressure Thrust method added

* BUG: call to the thrust function wrong

* BUG: pressure thrust evaluated when motor is turned off

* ENH: CHANGELOG updated

* DOC: definition of exhaust velocity improved

---------

Co-authored-by: Gui-FernandesBR <63590233+Gui-FernandesBR@users.noreply.github.com>