super() corrected and dict functions updated #801
The new parameter of the SolidMotor class was removed from super, since it's not on the Motor class. The dict functions were updated to take this new parameter into acount. Also the comments about the SolidMotor class parameters were updated.

Still need to do some tests running the code to be sure everything is ok, then I can open the PR.